refactor: tree provider & webview
