+++ extensions/iceworks-app/src/services.ts
@@ -1,0 +1,3 @@
+import * as common from '@iceworks/common-service';
+
+export default { common };
