+++ extensions/iceworks-app/src/views/npmScriptsView.ts
@@ -2,7 +2,8 @@ import * as vscode from 'vscode';
 import * as fse from 'fs-extra';
 import * as path from 'path';
 import { createNpmCommand } from '@iceworks/common-service'
-import { pathExists } from '../utils';
+import { pathExists, executeCommand } from '../utils';
+import { ITerminalMap } from '../types';
 
 export class NpmScriptsProvider implements vscode.TreeDataProvider<Script> {
   private workspaceRoot: string;
@@ -67,7 +68,7 @@ export class NpmScriptsProvider implements vscode.TreeDataProvider<Script> {
   }
 }
 
-export class Script extends vscode.TreeItem {
+class Script extends vscode.TreeItem {
   constructor(
     public readonly extensionContext: vscode.ExtensionContext,
     public readonly label: string,
@@ -83,4 +84,13 @@ export class Script extends vscode.TreeItem {
   };
 
   contextValue = 'script';
+}
+
+export function createNpmScriptsTreeProvider(context: vscode.ExtensionContext, rootPath: string, terminals: ITerminalMap) {
+  vscode.window.onDidCloseTerminal(term => terminals.delete(term.name));
+
+  const npmScriptsProvider = new NpmScriptsProvider(context, rootPath);
+  vscode.window.registerTreeDataProvider('npmScripts', npmScriptsProvider);
+  vscode.commands.registerCommand('iceworksApp.npmScripts.executeCommand', (script: Script) => executeCommand(terminals, script.command!));
+  vscode.commands.registerCommand('iceworksApp.npmScripts.refresh', () => npmScriptsProvider.refresh());
 }
